+++ ent/verifier/interface.go
@@ -20,5 +20,9 @@ type Verifier interface {
 	// MatchRecords function returns unsorted records corresponding to Input
 	// matches.  Matches contain an input name-string, and strings that matched
 	// that input.
-	MatchRecords(ctx context.Context, matches []mlib.Match) (map[string]*MatchRecord, error)
+	MatchRecords(
+		ctx context.Context,
+		matches []mlib.Match,
+		input vlib.Input,
+	) (map[string]*MatchRecord, error)
 }