+++ io/facetpg/facetpg.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"fmt"
 
 	"github.com/gnames/gnames/config"
 	"github.com/gnames/gnames/ent/facet"
@@ -24,7 +25,9 @@ type facetpg struct {
 }
 
 func New(cnf config.Config) facet.Facet {
-	db, err := sql.Open("postgres", dbshare.DBURL(cnf))
+	dbURL := dbshare.DBURL(cnf)
+	fmt.Printf("DB_URL: %#v\n\n", dbURL)
+	db, err := sql.Open("postgres", dbURL)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Cannot create PostgreSQL connection")
 	}