+++ io/rest/search_test.go
@@ -16,7 +16,7 @@ import (
 const searchURL = "http://:8888/api/v0/search"
 
 func TestGetSearch(t *testing.T) {
-	query := url.PathEscape("n:Bubo bubo tx:Aves ds:1,2 all:t au:Linn.")
+	query := url.PathEscape("n:Proh. wilsoni tx:Carnivora ds:172,11 all:t au:Gust.")
 	resp, err := http.Get(searchURL + "/" + query)
 	assert.Nil(t, err)
 	var response search.Output
@@ -27,10 +27,10 @@ func TestGetSearch(t *testing.T) {
 	assert.Nil(t, err)
 	meta := response.Meta
 	names := response.Names
-	assert.True(t, meta.Input.WithAllResults)
-	assert.Equal(t, []int{1, 2}, meta.Input.DataSourceIDs)
-	assert.Equal(t, "Linn.", meta.Input.Author)
-	assert.True(t, len(names) > 1)
+	assert.True(t, meta.Input.WithAllMatches)
+	assert.Equal(t, []int{172, 11}, meta.Input.DataSources)
+	assert.Equal(t, "Gust.", meta.Input.Author)
+	assert.True(t, len(names) > 0)
 	assert.True(t, len(names[0].Results) > 0)
 }
 
@@ -58,7 +58,7 @@ func TestPostSearch(t *testing.T) {
 	for _, v := range tests {
 		inp := gnquery.New().Parse(v.query)
 		assert.Equal(t, 0, len(inp.Warnings))
-		assert.False(t, inp.WithAllResults)
+		assert.False(t, inp.WithAllMatches)
 		req, err := gnfmt.GNjson{}.Encode(inp)
 		assert.Nil(t, err)
 		r := bytes.NewReader(req)